+++ Exp2_S4/src/main/java/com/mycompany/exp2_s4/ui/RegistrarUsuarioMenu.java
@@ -5,6 +5,12 @@ import com.mycompany.exp2_s4.servicio.ServiciosBiblioteca;
 import java.io.IOException;
 import java.util.Scanner;
 
+/*
+ * Submenu para registrar un nuevo usuario
+ * Este menú permite al usuario ingresar su ID, nombre, carrera y sede.
+ * Se valida la entrada y se maneja la excepción si el usuario ya existe.
+ * si el registro es exitoso, se guarda en disco.
+ */
 public class RegistrarUsuarioMenu {
 
     private final ServiciosBiblioteca svc;
@@ -16,7 +22,7 @@ public class RegistrarUsuarioMenu {
     }
 
     public void mostrar() {
-        System.out.println("\n--- Registrar Usuario ---");
+        System.out.println("\n----- Registrar Usuario -----");
         System.out.print("Ingrese su ID (RUT, sin puntos y con guion, ej. 19133090-0): ");
         String id = sc.nextLine().trim();
 
@@ -30,8 +36,10 @@ public class RegistrarUsuarioMenu {
             "ADEMINISTRACION DE EMPRESAS",
             "ADMINISTRACION PUBLICA",
             "INGENIERIA EN CONSTRUCCION",
-            "INGENIERIA CIVIL INDUSTRIAL",
+            "INGENIERIA CIVIL INDUSTRIAL"
         };
+        
+        // Mostrar las carreras disponibles
         System.out.println("\nSeleccione su carrera:");
         for (int i = 0; i < carreras.length; i++) {
             System.out.println((i + 1) + ". " + carreras[i]);
@@ -48,8 +56,10 @@ public class RegistrarUsuarioMenu {
             "SEDE MAIPU",
             "SEDE PUERTO MONTT",
             "SEDE ALAMEDA",
-            "CAMPUS VIRTUAL",
+            "CAMPUS VIRTUAL"
         };
+
+        // Mostrar las sedes disponibles
         System.out.println("\nSeleccione la sede:");
         for (int i = 0; i < sedes.length; i++) {
             System.out.println((i + 1) + ". " + sedes[i]);
@@ -61,6 +71,7 @@ public class RegistrarUsuarioMenu {
         }
         String sede = sedes[seleccionSede - 1];
 
+        // Intentar registrar el usuario
         try {
             svc.registrarUsuario(id, nombre, carrera, sede);
             System.out.println("Usuario registrado con éxito.");